requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
dotenv==0.9.9
idna==3.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
